add plus inc

+++ pageable/pagination.go
@@ -2,10 +2,10 @@ package pageable
 
 // Pagination http返回分页数据结构
 type Pagination struct {
-	Page       int   `json:"page"`
-	Size       int   `json:"size"`
-	Total      int64 `json:"total"`
-	TotalPages int   `json:"total_pages"`
+	Page       int `json:"page"`
+	Size       int `json:"size"`
+	Total      int `json:"total"`
+	TotalPages int `json:"total_pages"`
 }
 
 func NewPagination(total int64, page Pageable) *Pagination {
@@ -18,5 +18,5 @@ func NewPagination(total int64, page Pageable) *Pagination {
 	if mod > 0 && total > size {
 		totalPages += 1
 	}
-	return &Pagination{Total: total, TotalPages: int(totalPages), Page: page.Page(), Size: page.Size()}
+	return &Pagination{Total: int(total), TotalPages: int(totalPages), Page: page.Page(), Size: page.Size()}
 }
